+++ src/main/java/roomescape/payment/infrastructure/TossPaymentProperties.java
@@ -1,6 +1,7 @@
 package roomescape.payment.infrastructure;
 
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.context.properties.ConfigurationProperties;
@@ -18,4 +19,13 @@ public class TossPaymentProperties {
 
     @NotBlank(message = "widgetSecretKey은 비어있을 수 없습니다.")
     private final String widgetSecretKey;
+
+    @NotNull(message = "connectTimeout은 비어있을 수 없습니다.")
+    private final int connectTimeout;
+
+    @NotNull(message = "readTimeout은 비어있을 수 없습니다.")
+    private final int readTimeout;
+
+    @NotNull(message = "connectionRequestTimeout 비어있을 수 없습니다.")
+    private final int connectionRequestTimeout;
 }
