+++ src/main/resources/application.yml
@@ -23,13 +23,12 @@ security:
     token:
       secret-key: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIiLCJuYW1lIjoiSm9obiBEb2UiLCJpYXQiOjE1MTYyMzkwMjJ9.ih1aovtQShabQ7l0cINw4k1fagApg3qLWiB8Kt59Lno
       expire-length: 3600000
-rest-client:
-  connect-timeout: 2000
-  connection-request-timeout: 2000
-  read-timeout: 5000
 toss-payment:
   base-url: https://api.tosspayments.com
   widget-secret-key: test_gsk_docs_OaPz8L5KdmQXkzRz3y47BMw6
+  connect-timeout: 2000
+  connection-request-timeout: 2000
+  read-timeout: 5000
 logging:
   level:
     org.hibernate.SQL: debug
