+++ src/test/resources/application.yml
@@ -31,10 +31,9 @@ logging:
     org.hibernate.SQL: debug
     org.hibernate.orm.jdbc.bind: trace
 
-rest-client:
-  connect-timeout: 1000
-  connection-request-timeout: 1000
-  read-timeout: 2000
 toss-payment:
   widget-secret-key: test_gsk_docs_OaPz8L5KdmQXkzRz3y47BMw6
   base-url: http://localhost:8089
+  connect-timeout: 1000
+  connection-request-timeout: 1000
+  read-timeout: 2000
